requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
python-dotenv
psycopg2-binary
marshmallow
openai
